LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/meta-viewport-link-stylesheet-media.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"

    <!-- speculative case in document.write -->
    <meta name=viewport content="width=400, initial-scale=1"><link rel=stylesheet href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=9d80d37d-38b9-4b5b-80d7-a331e8a12739&amp;encodingcheck=&Gbreve;" media="(min-width: 401px)">


FAIL Speculative parsing, document.write(): meta-viewport-link-stylesheet-media Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\n"
